[platform/ios-sdk] FBLikeButton
Summary:
Initial commit of a Like button for native iOS apps.

Design still hasn't given us images, so for now I'm using some janky screenshots of the web's like button.

Test Plan:
Clicking "Like" causes a like to happen, and changes the button text to "Unlike". Clicking "Unlike" removes the like, and changes the button text back to "Like".

If you log in as User A, click Like, log out, and then log in as User B, the button reverts to the unliked state. If you log out and log back in as User B, the button shows itself as being liked again.

Reviewers: yariv, jimbru, nikolay, lshepard

Differential Revision: 247338