Add official app ID to the Getting Started demo
Summary:
Most people who download the app will want to just build, run, and see it
working. I think the extra hassle of configuring an app id makes it seem harder
than it is. If a developer does copy/paste this code and use it to bootstrap
their own app, great - they have plenty of opportunities to figure out that they
need to change the app id for their own, but I don't think the very first
experience is the right time for it.

Test Plan: Built app, ran through single sign on flow, was successful.

Reviewers: yariv, mattwkelly, caabernathy

Reviewed By: yariv

CC: brent, yariv

Differential Revision: 318879